// services/gateway/src/core/mongo/mongo.enums.ts
export enum MongoErrors {
    Conflict = 'conflict',
    NotFound = 'not-found',
    InvalidObjectId = 'invalid-object-id',
}

export enum MongoDataSource {
    Core = 'core-db',
    Office = 'office-db',
}

export enum ModelIdTag {
    Business = 'bz',
    BusinessHistory = 'bz.h',
    User = 'usr',
    Role = 'rol',
    Transaction = 'txn',
    AccountingGLCode = 'gl.c',
    AccountingGLMapping = 'gl.map',
    Budget = 'bud',
    Card = 'c',
    Account = 'ac',
    Approval = 'apr',
    CardOrder = 'c.ord',
    ScheduleTask = 'sch',
    PaymentSubscription = 'sub',
    Tag = 'tag',
    SpendControl = 'spc',
    Balance = 'bal',
    BalanceHistory = 'bal.h',
    OnboardSurvey = 'onb',
    Beneficiary = 'b',
    Audit = 'aud',
    BusinessKYC = 'bz.kyc',
    BusinessDirector = 'bz.dir',
    BusinessEntity = 'bz.ent',
    CardOrderHistory = 'c.ord.h',
    ShippingAddress = 'shp.adr',
    Merchant = 'mch',
    File = 'f',
    FileHistory = 'f.h',
    Team = 'tm',
    Depositor = 'dp',
}

export enum ObjectModel {
    Budget = 'Budget',
    User = 'User',
    Account = 'Account',
    Approval = 'Approval',
    CardOrder = 'CardOrder',
    Card = 'Card',
    Tag = 'Tag',
    Transaction = 'Transaction',
    ScheduleTask = 'ScheduleTask',
    Wallet = 'Wallet',
    SpendControl = 'SpendControl',
    Balance = 'Balance',
    BalanceHistory = 'BalanceHistory',
    Beneficiary = 'Beneficiary',
    Audit = 'Audit',
    Business = 'Business',
    BusinessHistory = 'BusinessHistory',
    Role = 'Role',
    AccountingGLCode = 'AccountingGLCode',
    AccountingGLMapping = 'AccountingGLMapping',
    PaymentSubscription = 'PaymentSubscription',
    OnboardSurvey = 'OnboardSurvey',
    BusinessKYC = 'BusinessKYC',
    CardOrderHistory = 'CardOrderHistory',
    ShippingAddress = 'ShippingAddress',
    Merchant = 'Merchant',
    File = 'File',
    FileHistory = 'FileHistory',
    Team = 'Team',
    BusinessKYCDirector = 'BusinessKYCDirector',
}